+++ projectHealthcalc/src/main/java/healthcalc/HealthCalc.java
@@ -2,7 +2,7 @@ package healthcalc;
 
 public interface HealthCalc {
 
-    public float getIdealBodyWeight(Person person) throws InvalidHeightException, InvalidGenderException;
+    public double getIdealBodyWeight(Person person) throws InvalidHeightException, InvalidGenderException;
 
-    public float basalMetabolicRate(Person person) throws InvalidValueException, InvalidGenderException;
+    public double basalMetabolicRate(Person person) throws InvalidValueException, InvalidGenderException;
 }
